move thymeleaf extension classes to bizfw
