+++ src/main/java/org/docksidestage/bizfw/thymeleaf/ExampleDialect.java
@@ -1,4 +1,4 @@
-package org.docksidestage.app.application;
+package org.docksidestage.bizfw.thymeleaf;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -8,11 +8,14 @@ import org.thymeleaf.processor.IProcessor;
 
 /**
  * @author subaru
+ * @author jflute
  */
 public class ExampleDialect extends AbstractProcessorDialect {
-    protected ExampleDialect(String name, String prefix, int processorPrecedence) {
+
+    public ExampleDialect(String name, String prefix, int processorPrecedence) {
         super(name, prefix, processorPrecedence);
     }
+
     @Override
     public Set<IProcessor> getProcessors(String dialectPrefix) {
         HashSet<IProcessor> processors = new HashSet<>();
